partitioned coral intake/wrist and algae intake/wrist
