+++ src/main/cpp/subsystems/CoralIntake.cpp
@@ -4,8 +4,114 @@ namespace t34
 {    
   CoralIntake::CoralIntake()
     : m_motor(4, rev::spark::SparkLowLevel::MotorType::kBrushless)
+    , m_coral_level(0)
+    , m_init_coral_angle(0_deg) 
+    , m_coral_wrist_motor(3, SparkLowLevel::MotorType::kBrushless)
+    , m_coral_wrist_pid(0.25, 0.0, 0.0)
   {
+    m_coral_wrist_pid.SetTolerance( (0.25_deg) / 1_tr );
   } 
+
+  frc2::CommandPtr CoralIntake::MoveCoralWristByPowerCommand(double val)
+  {
+    return this->RunEnd
+    (
+      [this, val] 
+      { 
+        m_coral_wrist_motor.Set(-val);  
+      },
+      [this]
+      {
+        m_coral_wrist_motor.Set(0.0);
+      }
+    );
+  }
+
+   frc2::CommandPtr CoralIntake::MoveToCoralLevelCommand(int level)
+  {
+    static const std::array<double, 5> presets {{
+      0.0, 4.0, 13.666, 19.83, 23.0
+    }};
+
+    m_coral_level = std::clamp(level, 0, static_cast<int>(presets.size()) - 1);
+
+    // m_coral_level = std::clamp(level, 0, 4);
+
+    // switch (m_coral_level)
+    // {
+    //   case 0:
+    //     return MoveCoralWristToCommand(0.0);
+    //     break;
+    //   case 1:
+    //     return MoveCoralWristToCommand(4.0);
+    //     break;
+    // }
+
+    return MoveCoralWristToCommand(presets.at(m_coral_level));
+  }
+
+   frc2::CommandPtr CoralIntake::IncrementCoralDown()
+  {
+    return this->MoveToCoralLevelCommand(m_coral_level + 1);  
+  }
+
+  frc2::CommandPtr CoralIntake::IncrementCoralUp()
+  {
+    return this->MoveToCoralLevelCommand(m_coral_level - 1); 
+  }
+
+
+ frc2::CommandPtr CoralIntake::MoveCoralWristToCommand(double encoder_units)
+  {
+
+    bool run_up = m_coral_wrist_motor.GetEncoder().GetPosition() < encoder_units;
+    //m_coral_wrist_pid.SetSetpoint(encoder_units);//-(angle.value() - m_init_algae_angle.value()) * 11.923);//BOTH_WRIST_GEAR_RATIO * ((angle - m_init_coral_angle) / 1_tr));
+
+
+
+    return this->RunEnd(
+      [this, run_up]
+      {
+        //m_coral_wrist_motor.Set(m_coral_wrist_pid.Calculate(m_coral_wrist_motor.GetEncoder().GetPosition()));
+
+        m_coral_wrist_motor.Set((run_up) ? 0.3 : -0.3);
+      },
+      [this]
+      {
+          m_coral_wrist_motor.StopMotor();
+      })
+      .Until([this, encoder_units, run_up] 
+      { 
+        if (run_up)
+        {
+          return m_coral_wrist_motor.GetEncoder().GetPosition() >= (encoder_units - 0.1);
+        }
+        else
+        {
+          return m_coral_wrist_motor.GetEncoder().GetPosition() <= (encoder_units + 0.1);
+        }
+      });
+  }
+
+frc2::CommandPtr CoralIntake::MoveCoralWristToCommand(units::degree_t angle)
+  {
+    std::clamp(angle, 0_deg, 150_deg);
+    m_coral_wrist_pid.SetSetpoint(-426.755 + (0.0438 * (angle.value() - m_init_coral_angle.value())));//-(angle.value() - m_init_algae_angle.value()) * 11.923);//BOTH_WRIST_GEAR_RATIO * ((angle - m_init_coral_angle) / 1_tr));
+
+    return this->RunEnd(
+      [this, angle]
+      {
+        m_coral_wrist_motor.Set(m_coral_wrist_pid.Calculate(m_coral_wrist_motor.GetEncoder().GetPosition()));
+      },
+      [this]
+      {
+          m_coral_wrist_motor.StopMotor();
+      })
+      .Until([this] 
+      { 
+        return m_coral_wrist_pid.AtSetpoint();
+      });
+  }
 
   frc2::CommandPtr CoralIntake::RunInCommand()
   {
@@ -21,5 +127,12 @@ namespace t34
       [this] { this->m_motor.Set(0.5); },
       [this] { this->m_motor.StopMotor(); }
     );
+  }
+
+ void CoralIntake::Periodic()
+  {
+    frc::SmartDashboard::PutNumber("Coral Wrist Encoder: ", m_coral_wrist_motor.GetEncoder().GetPosition());
+    frc::SmartDashboard::PutNumber("Coral Setpoint", m_coral_wrist_pid.GetSetpoint());
+
   }
 }
