+++ src/main/cpp/subsystems/Elevator.cpp
@@ -14,137 +14,14 @@ namespace t34
 
   Elevator::Elevator()
   : m_level(0)
-  , m_coral_level(0)
-
   , m_init_height(16_in)
-
     //The wrists' angles are from 0 to 180 degrees (0 is straight down, 180 is straight up, and 90 is parallel to the floor)
-  , m_init_algae_angle(155_deg) //65 degrees away from horizontal
-  , m_init_coral_angle(0_deg) 
-
-  , m_right_algae_wrist_motor(1)
-  , m_left_algae_wrist_motor(2)
-
   , m_left_motor(11)
   , m_right_motor(12)
-
-  , m_coral_wrist_motor(3, SparkLowLevel::MotorType::kBrushless)
-
   , m_elevator_motors_pid(0.5, 0.0, 0.0)
-  , m_algae_wrist_pid(0.5, 0.0, 0.0)
-  , m_coral_wrist_pid(0.25, 0.0, 0.0)
   {
     m_elevator_motors_pid.SetTolerance(Neo::LengthTo550Unit(0.5_in));
-    m_coral_wrist_pid.SetTolerance( (0.25_deg) / 1_tr );
-    m_algae_wrist_pid.SetTolerance( (0.25_deg) / 1_tr);
   }
-
-  frc2::CommandPtr Elevator::MoveAlgaeWristToCommand(double encoder_units)
-  {                                             //units::degree_t angle
-    //std::clamp(angle, 0_deg, 115_deg);
-    
-    double clamped_encoder_units = std::clamp(encoder_units, -9000.0, 593.0);
-    //m_algae_wrist_pid.SetSetpoint(BOTH_WRIST_GEAR_RATIO * ((angle - m_init_algae_angle) / 1_tr));
-    m_algae_wrist_pid.SetSetpoint(clamped_encoder_units);
-                      //BOTH_WRIST_GEAR_RATIO * ((angle - m_init_algae_angle) / 1_tr)
-
-    return this->RunEnd(
-      [this, clamped_encoder_units]
-      {
-        m_left_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::Position, m_algae_wrist_pid.Calculate(m_left_algae_wrist_motor.GetSelectedSensorPosition()));
-        m_right_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::Position, m_algae_wrist_pid.Calculate(m_right_algae_wrist_motor.GetSelectedSensorPosition()));
-      },
-      [this]
-      {
-        m_left_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0);
-        m_right_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0);
-      })
-      .Until([this] 
-      { 
-        return m_algae_wrist_pid.AtSetpoint();
-      });
-  }
-
-  frc2::CommandPtr Elevator::MoveAlgaeWristByIncrementCommand(double increase) {
-    
-    return this->RunEnd(
-      [this, increase]
-      {
-        m_right_algae_wrist_motor.Set(
-          ctre::phoenix::motorcontrol::ControlMode::PercentOutput,
-          m_algae_wrist_pid.Calculate(
-            m_right_algae_wrist_motor.GetSelectedSensorPosition(), 
-            (m_algae_wrist_pid.GetSetpoint() + increase)));
-
-        m_left_algae_wrist_motor.Set(
-        ctre::phoenix::motorcontrol::ControlMode::PercentOutput,
-          m_algae_wrist_pid.Calculate(
-            m_left_algae_wrist_motor.GetSelectedSensorPosition(), 
-            (m_algae_wrist_pid.GetSetpoint() + increase)));
-      },
-      [this]
-      {
-          m_left_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0);
-          m_right_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0);
-      })
-      .Until([this] 
-      { 
-        return m_algae_wrist_pid.AtSetpoint();
-      });
-  }
-
-frc2::CommandPtr Elevator::MoveCoralWristToCommand(units::degree_t angle)
-  {
-    std::clamp(angle, 0_deg, 150_deg);
-    m_coral_wrist_pid.SetSetpoint(-426.755 + (0.0438 * (angle.value() - m_init_algae_angle.value())));//-(angle.value() - m_init_algae_angle.value()) * 11.923);//BOTH_WRIST_GEAR_RATIO * ((angle - m_init_coral_angle) / 1_tr));
-
-    return this->RunEnd(
-      [this, angle]
-      {
-        m_coral_wrist_motor.Set(m_coral_wrist_pid.Calculate(m_coral_wrist_motor.GetEncoder().GetPosition()));
-      },
-      [this]
-      {
-          m_coral_wrist_motor.StopMotor();
-      })
-      .Until([this] 
-      { 
-        return m_coral_wrist_pid.AtSetpoint();
-      });
-  }
-
-  frc2::CommandPtr Elevator::MoveCoralWristToCommand(double encoder_units)
-  {
-
-    bool run_up = m_coral_wrist_motor.GetEncoder().GetPosition() < encoder_units;
-    //m_coral_wrist_pid.SetSetpoint(encoder_units);//-(angle.value() - m_init_algae_angle.value()) * 11.923);//BOTH_WRIST_GEAR_RATIO * ((angle - m_init_coral_angle) / 1_tr));
-
-
-
-    return this->RunEnd(
-      [this, run_up]
-      {
-        //m_coral_wrist_motor.Set(m_coral_wrist_pid.Calculate(m_coral_wrist_motor.GetEncoder().GetPosition()));
-
-        m_coral_wrist_motor.Set((run_up) ? 0.3 : -0.3);
-      },
-      [this]
-      {
-          m_coral_wrist_motor.StopMotor();
-      })
-      .Until([this, encoder_units, run_up] 
-      { 
-        if (run_up)
-        {
-          return m_coral_wrist_motor.GetEncoder().GetPosition() >= (encoder_units - 0.1);
-        }
-        else
-        {
-          return m_coral_wrist_motor.GetEncoder().GetPosition() <= (encoder_units + 0.1);
-        }
-      });
-  }
-
 
   frc2::CommandPtr Elevator::ElevateToCommand(units::inch_t height)
   {
@@ -184,7 +61,7 @@ frc2::CommandPtr Elevator::MoveCoralWristToCommand(units::degree_t angle)
     
     auto [angle, height] = presets.at(level);
 
-    return ElevateToCommand(height).AndThen(MoveCoralWristToCommand(angle));
+    return ElevateToCommand(height).AndThen(m_coral_intake.MoveCoralWristToCommand(angle));
   };
 
   frc2::CommandPtr Elevator::MoveUpOnceCommand()
@@ -220,77 +97,8 @@ frc2::CommandPtr Elevator::MoveCoralWristToCommand(units::degree_t angle)
     );
   }
 
-  frc2::CommandPtr Elevator::MoveAlgaeWristByPowerCommand(double val)
-  {
-    return this->RunEnd
-    (
-      [this, val] 
-        { 
-          m_left_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, val);
-          m_right_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, -val); 
-        },
-        [this]
-        {
-          m_left_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0);
-          m_right_algae_wrist_motor.Set(ctre::phoenix::motorcontrol::ControlMode::PercentOutput, 0.0); 
-        }
-    );
-  }
-
-  frc2::CommandPtr Elevator::MoveCoralWristByPowerCommand(double val)
-  {
-    return this->RunEnd
-    (
-      [this, val] 
-      { 
-        m_coral_wrist_motor.Set(-val);  
-      },
-      [this]
-      {
-        m_coral_wrist_motor.Set(0.0);
-      }
-    );
-  }
-
-  frc2::CommandPtr Elevator::MoveToCoralLevelCommand(int level)
-  {
-    static const std::array<double, 5> presets {{
-      0.0, 4.0, 13.666, 19.83, 23.0
-    }};
-
-    m_coral_level = std::clamp(level, 0, static_cast<int>(presets.size()) - 1);
-
-    // m_coral_level = std::clamp(level, 0, 4);
-
-    // switch (m_coral_level)
-    // {
-    //   case 0:
-    //     return MoveCoralWristToCommand(0.0);
-    //     break;
-    //   case 1:
-    //     return MoveCoralWristToCommand(4.0);
-    //     break;
-    // }
-
-    return MoveCoralWristToCommand(presets.at(m_coral_level));
-  }
-
-  frc2::CommandPtr Elevator::IncrementCoralDown()
-  {
-    return this->MoveToCoralLevelCommand(m_coral_level + 1);  
-  }
-
-  frc2::CommandPtr Elevator::IncrementCoralUp()
-  {
-    return this->MoveToCoralLevelCommand(m_coral_level - 1); 
-  }
 
 
-  void Elevator::Periodic() {
-      frc::SmartDashboard::PutNumber("Left Algae Wrist Encoder", m_left_algae_wrist_motor.GetSelectedSensorPosition());
-      frc::SmartDashboard::PutNumber("Right Algae Wrist Encoder", m_right_algae_wrist_motor.GetSelectedSensorPosition());
-      frc::SmartDashboard::PutNumber("Coral Wrist Encoder: ", m_coral_wrist_motor.GetEncoder().GetPosition());
-      frc::SmartDashboard::PutNumber("Coral Setpoint", m_coral_wrist_pid.GetSetpoint());
-  }
+     
 
 } // namespace t34
