+++ src/main/include/subsystems/AlgaeIntake.h
@@ -5,6 +5,17 @@
 #include <frc2/command/InstantCommand.h>
 #include <rev/config/SparkMaxConfig.h>
 #include "Intake.h"
+#include <frc/smartdashboard/SmartDashboard.h>
+#include <ctre/phoenix/motorcontrol/can/TalonSRX.h>
+#include <frc/controller/PIDController.h>
+#include <string>
+#include <cmath>
+#include <units/length.h>
+#include <units/angle.h>
+
+using namespace ctre::phoenix6::hardware;
+using namespace rev::spark;
+using namespace ctre::phoenix::motorcontrol::can;
 
 namespace t34
 {
@@ -15,8 +26,23 @@ namespace t34
 
         virtual frc2::CommandPtr RunInCommand() override;
         virtual frc2::CommandPtr RunOutCommand() override;
+
+        void Periodic() override;
+
+        frc2::CommandPtr MoveAlgaeWristToCommand(double enc_units); //units::degree_t angle
+        frc2::CommandPtr MoveAlgaeWristByPowerCommand(double val);
+        frc2::CommandPtr MoveAlgaeWristByIncrementCommand(double increase);
+
         
     private:
         ctre::phoenix6::hardware::TalonFX m_motor;
+
+        const units::degree_t m_init_algae_angle;
+
+        TalonSRX m_right_algae_wrist_motor;
+        TalonSRX m_left_algae_wrist_motor;
+
+        frc::PIDController m_algae_wrist_pid;
+        
     };
-}
+};
